[Speculative Decoding] Add ProposerWorkerBase abstract class
SpecDecodeWorker and Top1Proposer make assumptions about some methods that exist on proposer_worker but these aren't captured by the typing.
